Refactor to injectFiles
