+++ src/content/addChangedFiles.tsx
@@ -1,9 +1,7 @@
 import { Octokit } from "@octokit/rest";
-import React from "react";
-import { createRoot } from "react-dom/client";
 import { Settings } from "../services";
+import { injectFiles } from "./injectPrFiles";
 import { injectPrFilesSearch } from "./injectPrFilesSearch";
-import { PrFiles } from "./PrFiles";
 import { processPrFiles } from "./processPrFiles";
 import { Files } from "./types";
 
@@ -30,32 +28,5 @@ export async function addChangedFiles(settings: Settings) {
   }
 
   injectPrFilesSearch(prs, prFilesMap);
-
-  // Add Files-Icon to PRs
-  const prRows = document.querySelectorAll("div[id^=issue_]");
-  for await (const prRow of prRows) {
-    // Add labels indicating the base-ref to every PR row
-    const [, prNumber] = prRow.id.split("_");
-    //  Add popover for the files
-    const prFiles = prFilesMap.get(parseInt(prNumber));
-    // addPopover(prFiles, prRow);
-    const prFilesClass = "gh-ui-booster-pr-files";
-    if (prRow.classList.contains(prFilesClass)) return;
-    prRow.classList.add(prFilesClass);
-
-    const rootSpanEl = document.createElement("span");
-    // Insert between the first and second child
-    prRow.children[0].children[2].insertBefore(
-      rootSpanEl,
-      prRow.children[0].children[2].children[1]
-    );
-
-    const root = createRoot(rootSpanEl);
-
-    root.render(
-      <React.StrictMode>
-        <PrFiles prFiles={prFiles} />
-      </React.StrictMode>
-    );
-  }
+  injectFiles(prFilesMap);
 }
